requirements.txt
python-dotenv
google-genai
fastapi
uvicorn
a2a-sdk
httpx
cryptography
